requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
requests==2.31.0
pytest==8.0.0
httpx==0.26.0
google-generativeai==0.3.2